move parameter fixing to another file
